chore: usersExperiences updated with experiences

+++ src/modules/experiences/services/experiences.service.ts
@@ -3,12 +3,15 @@ import { ExperiencesRepository } from "../repositories/experiences.repository";
 import { FindAllDto } from "@/modules/common/dto/find-all.dto";
 import { CreateExperienceDto } from "@/modules/experiences/dto/create-experience.dto";
 import { UpdateExperienceDto } from "@/modules/experiences/dto/update-experience.dto";
+import { UsersSkillsService } from "@/modules/users/services/users-skills.service";
 
 class ExperiencesService {
   private readonly experiencesRepository: ExperiencesRepository;
+  private readonly usersSkillsService: UsersSkillsService;
 
   constructor() {
     this.experiencesRepository = new ExperiencesRepository();
+    this.usersSkillsService = new UsersSkillsService();
   }
 
   findAll(findAllDto: FindAllDto) {
@@ -19,16 +22,43 @@ class ExperiencesService {
     return this.experiencesRepository.findOne(id);
   }
 
-  create(createExperienceDto: CreateExperienceDto, userId: string) {
-    return this.experiencesRepository.create(createExperienceDto, userId);
+  async create(createExperienceDto: CreateExperienceDto, userId: string) {
+    const experience = await this.experiencesRepository.create(
+      createExperienceDto,
+      userId
+    );
+    await this.usersSkillsService.updateUserSkillsFromExperience(
+      userId,
+      experience.id
+    );
+    return experience;
   }
 
-  update(id: string, updateExperienceDto: UpdateExperienceDto) {
-    return this.experiencesRepository.update(id, updateExperienceDto);
+  async update(id: string, updateExperienceDto: UpdateExperienceDto) {
+    const experience = await this.experiencesRepository.findOne(id);
+    if (!experience) throw new Error("Experience not found");
+
+    const updatedExperience = await this.experiencesRepository.update(
+      id,
+      updateExperienceDto
+    );
+    await this.usersSkillsService.updateUserSkillsFromExperience(
+      experience.userId,
+      id
+    );
+    return updatedExperience;
   }
 
-  delete(id: string) {
-    return this.experiencesRepository.delete(id);
+  async delete(id: string) {
+    const experience = await this.experiencesRepository.findOne(id);
+    if (!experience) throw new Error("Experience not found");
+
+    const deletedExperience = await this.experiencesRepository.delete(id);
+    await this.usersSkillsService.updateUserSkillsFromExperience(
+      experience.userId,
+      id
+    );
+    return deletedExperience;
   }
 }
